feat: Added code for Role based access

+++ src/Apha.BST/Apha.BST.Web/Controllers/SiteController.cs
@@ -3,10 +3,13 @@ using Apha.BST.Application.DTOs;
 using Apha.BST.Application.Interfaces;
 using Apha.BST.Core.Entities;
 using Apha.BST.Web.Models;
+using Apha.BST.Web.PresentationService;
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.Data.SqlClient;
 using static Apha.BST.Web.Models.SiteViewModel;
 
 namespace Apha.BST.Web.Controllers
@@ -15,22 +18,25 @@ namespace Apha.BST.Web.Controllers
     public class SiteController : Controller
     {
         private readonly ISiteService _siteService;
+        private readonly IUserDataService _userDataService;
+        private readonly ILogger<SiteController> _logger;
         private readonly IMapper _mapper;
         private const string siteAll = "All";       
 
-        public SiteController(ISiteService siteService, IMapper mapper)
+        public SiteController(ILogger<SiteController> logger,ISiteService siteService, IMapper mapper,IUserDataService userDataService)
         {
             _siteService = siteService;
-            _mapper = mapper;          
-        }
-        public IActionResult Index()
-        {
-            return View();
-        }
+            _mapper = mapper;
+            _userDataService = userDataService;
+            _logger = logger;
+        }       
 
         [HttpGet]
         public async Task<IActionResult> ViewSite(string selectedSite = siteAll)
         {
+            bool canEdit = await _userDataService.CanEditPage(ControllerContext.ActionDescriptor.ActionName);
+           
+
             var allSitesDto = await _siteService.GetAllSitesAsync("All");
             var allSites = _mapper.Map<IEnumerable<SiteViewModel>>(allSitesDto);
 
@@ -50,7 +56,8 @@ namespace Apha.BST.Web.Controllers
             {
                 AllSites = siteSelectList,
                 FilteredSites = filteredSites,
-                SelectedSite = selectedSite
+                SelectedSite = selectedSite,
+                CanEdit = canEdit
             };
 
             return View(model);
@@ -59,6 +66,7 @@ namespace Apha.BST.Web.Controllers
         //Final code working for SiteTrainee
         public async Task<IActionResult> SiteTrainee(string? selectedSite = null)
         {
+            bool canEdit = await _userDataService.CanEditPage(ControllerContext.ActionDescriptor.ActionName);
             // Get all sites for the dropdown
             var allSitesDto = await _siteService.GetAllSitesAsync(siteAll);
             var allSites = _mapper.Map<IEnumerable<SiteViewModel>>(allSitesDto);
@@ -75,7 +83,8 @@ namespace Apha.BST.Web.Controllers
             var model = new SiteTraineeListViewModel
             {
 
-                AllSites = siteSelectList
+                AllSites = siteSelectList,
+                CanEdit = canEdit
             };
             if (selectedSite != null)
             {
@@ -87,6 +96,7 @@ namespace Apha.BST.Web.Controllers
 
                 model.FilteredTrainees = filteredTrainees;
                 model.SelectedSite = selectedSite;
+               
 
             }
 
@@ -96,28 +106,48 @@ namespace Apha.BST.Web.Controllers
 
 
         [HttpGet]
-        public IActionResult AddSite()
+        public async Task<IActionResult> AddSite()
         {
-            return View();
+            bool canEdit = await _userDataService.CanEditPage(ControllerContext.ActionDescriptor.ActionName);
+            SiteViewModel model = new SiteViewModel
+            {
+                PlantNo = string.Empty,
+                Name = string.Empty,
+                CanEdit = canEdit, // Assuming the user can edit when adding a site
+            };
+            return View(model);
         }
         
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> AddSite(SiteViewModel siteViewModel)
         {
+            bool canEdit = await _userDataService.CanEditPage(ControllerContext.ActionDescriptor.ActionName);
             if (ModelState.IsValid)
             {
                 try
                 {
-                    var site = _mapper.Map<SiteDto>(siteViewModel);
-                    var message = await _siteService.AddSiteAsync(site);
+                    string userName = _userDataService.GetUsername() ?? "";
+                    if (canEdit)
+                    {
+                        var site = _mapper.Map<SiteDto>(siteViewModel);
+                        var message = await _siteService.AddSiteAsync(site,userName);
 
-                    TempData["Message"] = message;
+                        TempData["SiteMessage"] = message;
+                    }
                 }
-                catch (Exception)
+                catch (SqlException sqlEx)
                 {
-                    TempData["Message"] = "Save failed";
+                    // Log SQL Exception with identifier (CloudWatch will receive this)
+                    _logger.LogError(sqlEx, "[BST.SQLException] Error in [AddSite]: {Message}", sqlEx.Message);
+                    TempData["SiteMessage"] = "Save failed";
                 }
+                catch (Exception ex)
+                {
+                    _logger.LogError(ex, "[BST.GENERAL_EXCEPTION] Error in [AddSite]: {Message}", ex.Message);
+                    TempData["SiteMessage"] = "Save failed";
+                }
+               
 
                 return RedirectToAction(nameof(AddSite));
             }
@@ -128,22 +158,34 @@ namespace Apha.BST.Web.Controllers
         [HttpPost]
         public async Task<IActionResult> DeleteTrainee(int personId, string selectedSite)
         {
-            if(!ModelState.IsValid)
+            bool canEdit = await _userDataService.CanEditPage(ControllerContext.ActionDescriptor.ActionName);
+            if (!ModelState.IsValid)
             {
-                TempData["message"] = "Invalid request";
+                TempData["SiteMessage"] = "Invalid request";
                 return RedirectToAction("SiteTrainee", new { selectedSite });
             }
             try
             {
-                var message = await _siteService.DeleteTraineeAsync(personId);
-                TempData["message"] = message;
+                if (canEdit)
+                {
+                    var message = await _siteService.DeleteTraineeAsync(personId);
+                    TempData["SiteMessage"] = message;
+                }
             }
-            catch (Exception)
+            catch (SqlException sqlEx)
             {
-                TempData["message"] = "Save failed";
+                // Log SQL Exception with identifier (CloudWatch will receive this)
+                _logger.LogError(sqlEx, "[BST.SQLException] Error in [DeleteTrainne]: {Message}", sqlEx.Message);
+                TempData["SiteMessage"] = "Save failed";
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, "[BST.GENERAL_EXCEPTION] Error in [DeleteTrainee]: {Message}", ex.Message);
+                TempData["SiteMessage"] = "Save failed";
             }
 
             return RedirectToAction("SiteTrainee", new { selectedSite });
         }
+       
     }
 }
